+++ src/briefy/common/users.py
@@ -1,0 +1,16 @@
+"""System users."""
+from briefy.common.types import BaseUser
+
+
+SystemUser = BaseUser(
+    user_id='be319e15-d256-4587-a871-c3476affa309',
+    data={
+        'locale': 'en_GB',
+        'fullname': 'Briefy',
+        'first_name': 'Briefy',
+        'last_name': '',
+        'email': 'app@briefy.co',
+        'groups': ['g:system', ],
+    }
+)
+